+++ ESFramework/UIKit/UIDevice+ESInfo.h
@@ -39,12 +39,36 @@ NS_ASSUME_NONNULL_BEGIN
 - (nullable NSString *)carrierName;
 
 /**
+ * Returns the current WiFi network info.
+ *
+ * @code
+ * {
+ *     BSSID = "20:c9:d0:e1:78:c9";
+ *     SSID = "Elf Sundae's MBP";
+ *     SSIDDATA = <456c6620 53756e64 61652773 204d4250>;
+ * }
+ * @endcode
+ *
+ * @warning To use this function in iOS 12 and later, enable the Access WiFi Information capability in Xcode.
+ * For more information, see https://developer.apple.com/documentation/bundleresources/entitlements/com_apple_developer_networking_wifi-info
+ */
+- (nullable NSDictionary *)WiFiNetworkInfo;
+
+/**
  * Returns the current WiFi SSID.
  *
  * @warning To use this function in iOS 12 and later, enable the Access WiFi Information capability in Xcode.
  * For more information, see https://developer.apple.com/documentation/bundleresources/entitlements/com_apple_developer_networking_wifi-info
  */
 - (nullable NSString *)WiFiSSID;
+
+/**
+ * Returns the current WiFi BSSID.
+ *
+ * @warning To use this function in iOS 12 and later, enable the Access WiFi Information capability in Xcode.
+ * For more information, see https://developer.apple.com/documentation/bundleresources/entitlements/com_apple_developer_networking_wifi-info
+ */
+- (nullable NSString *)WiFiBSSID;
 
 /**
  * Detects whether this device has been jailbroken.